requirements.txt
docopts
icalendar
psycopg2-binary
python-dateutil
python-dotenv
pyyaml==5.4
recurrent
recurring-ical-events
requests
schema
